requirements.txt
flask
aiogram
python-dotenv
#sqlite3  # Built-in, optional
stripe
paystackease
requests